2020-05-18 AC: Fixes an issue with clean restarts on local/Windows computers.

+++ src/Command/DockerManageCommand.php
@@ -688,14 +688,14 @@ class DockerManageCommand extends Command
 
                 // @codeCoverageIgnoreStart
                 if (!empty($this->tempScriptFile)) {
-                    // Allow for a clean restart with Docker.
-                    if (strtoupper(substr(PHP_OS, 0, 3)) !== 'WIN') {
+                    // Allow for a clean restart with Docker on Cloud servers using a Linux/Unix OS.
+                    if (LFPHP && strtoupper(substr(PHP_OS, 0, 3)) !== 'WIN') {
                         unlink($this->tempScriptFile);
+
+                        file_put_contents($this->tempScriptFile, "#!/usr/bin/env bash\nlfphp");
+
+                        chmod($this->tempScriptFile, 777); // Must be world-writable for Mac computers.
                     }
-
-                    file_put_contents($this->tempScriptFile, "#!/usr/bin/env bash\nlfphp");
-
-                    chmod($this->tempScriptFile, 777); // Must be world-writable for Mac computers.
                 }
                 // @codeCoverageIgnoreEnd
